requirements.txt
# dependencies used in the production environment.
# pip freeze = to see de dependencies in the environment
# pip install -r requirements.txt = to install the dependencies in the environment
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.0
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
jwcrypto==1.5.6
mysql-connector-python==9.1.0
mysqlclient==2.2.7
psycopg2==2.9.10
pycparser==2.22
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1